# src/main/resources/static/swagger.yaml
openapi: 3.0.1
servers:
  - url: /
info:
  title: Example Project Event Sourcing
  version: 1.0.1-SNAPSHOT
  description: Share Your Experience @LHIND, 2021-09-24

paths:

  /auftrag:
    post:
      summary: Erstelle Auftrag
      description: Einen neuen Auftrag erstellen.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuftragData'
      responses:
        201:
          description: Created

  /auftrag/{id}:
    get:
      summary: Erhalte Auftrag
      description: Erhalte die Read-Repräsentanz eines Auftags.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        200:
          description: Read-Modell des Auftrags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuftragCurrentState'
        404:
          $ref: '#/components/responses/404'
    put:
      summary: Einen Auftrag aktualisieren
      description: Aktualisiere einen bestehenden Auftrag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuftragData'
      responses:
        204:
          description: No Conent
        404:
          $ref: '#/components/responses/404'

  /auftrag/{id}/position:
    post:
      summary: Position hinzufügen
      description: Füge dem Auftrag eine Position hinzu.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FuegePositionHinzuCommand'
      responses:
        201:
          description: Created

  /auftrag/{id}/position/{positionId}:
    delete:
      summary: Position löschen
      description: Lösche eine Position in einem Auftrag.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: positionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        204:
          description: No Content
        404:
          $ref: '#/components/responses/404'


components:

  schemas:

    UUID:
      description: Ein eindeutiger Identifikator
      type: string
      format: uuid
      example: 0fc46152-83ca-489c-a2a3-6dc1664517db

    AuftragData:
      type: object
      properties:
        beladestelle:
          $ref: '#/components/schemas/Ladestelle'
        entladestelle:
          $ref: '#/components/schemas/Ladestelle'

    FuegePositionHinzuCommand:
      type: object
      properties:
        bezeichnung:
          type: string
          example: Haushaltswaren
        warenwert:
          type: string
          pattern: '^\d+\.\d{2}$'
          example: '1000.50'

    Ladestelle:
      type: object
      properties:
        plz:
          type: string
          example: 27283
        ladezeit:
          type: string
          format: date-time

    Position:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        bezeichnung:
          type: string
          example: Haushaltswaren
        warenwert:
          type: string
          pattern: '^\d+\.\d{2}$'
          example: '1000.50'

    AuftragCurrentState:
      type: object
      properties:
        beladestelle:
          $ref: '#/components/schemas/Ladestelle'
        entladestelle:
          $ref: '#/components/schemas/Ladestelle'
        positionen:
          type: array
          items:
            $ref: '#/components/schemas/Position'

    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: number
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: 'Resource id=0fc46152-83ca-489c-a2a3-6dc1664517db nicht gefunden'
        path:
          type: string
          example: '/auftrag/0fc46152-83ca-489c-a2a3-6dc1664517db'

  responses:

    404:
      description: Die Ressource konnte nicht gefunden werden.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'